Update package to Laravel 5.4
From the Laravel official documentation : 
https://laravel.com/docs/5.4/upgrade
share Method Removed

The share method has been removed from the container. This was a legacy method that has not been documented in several years. If you are using this method, you should begin using the singleton method instead:

```
$container->singleton('foo', function () {
    return 'foo';
});
```

This commit refactors the AdministratorServiceProvider.php to use singletons instead of the now deprecated share.